KAM-89: [FE] Internationalization KAM-90: [FE] Internationalization `Authorize` module Refactored some code
