+++ src/app/pages/authorize/pages/sign-up/sign-up.module.ts
@@ -8,6 +8,7 @@ import { ButtonModule } from '@widgets/button';
 import { NgForOf } from '@angular/common';
 import { NzFormModule } from 'ng-zorro-antd/form';
 import { NzIconModule } from 'ng-zorro-antd/icon';
+import { TranslateModule } from '@ngx-translate/core';
 
 @NgModule({
   imports: [
@@ -19,6 +20,7 @@ import { NzIconModule } from 'ng-zorro-antd/icon';
     NgForOf,
     NzFormModule,
     NzIconModule,
+    TranslateModule,
   ],
   declarations: [SignUpComponent],
 })
