--- src/app/utils/constants/index.ts
@@ -1,4 +1,0 @@
-export * from './color.constant';
-export * from './form.constant';
-export * from './locale.constant';
-export * from './country-code.constant';
